+++ src/datasource-zabbix/metricFunctions.js
@@ -8,7 +8,8 @@ var categories = {
   Filter: [],
   Trends: [],
   Time: [],
-  Alias: []
+  Alias: [],
+  Special: []
 };
 
 function addFuncDef(funcDef) {
@@ -220,6 +221,16 @@ addFuncDef({
     { name: 'newAlias', type: 'string' }
   ],
   defaultParams: ['/(.*)/', '$1']
+});
+
+// Special
+addFuncDef({
+  name: 'consolidateBy',
+  category: 'Special',
+  params: [
+    { name: 'type', type: 'string', options: ['avg', 'min', 'max', 'sum', 'count'] }
+  ],
+  defaultParams: ['avg'],
 });
 
 _.each(categories, function(funcList, catName) {