+++ src/datasource-zabbix/module.js
@@ -1,8 +1,6 @@
 import {ZabbixAPIDatasource} from './datasource';
 import {ZabbixQueryController} from './query.controller';
-
-class ZabbixConfigController {}
-ZabbixConfigController.templateUrl = 'datasource-zabbix/partials/config.html';
+import {ZabbixDSConfigController} from './config.controller';
 
 class ZabbixQueryOptionsController {}
 ZabbixQueryOptionsController.templateUrl = 'datasource-zabbix/partials/query.options.html';
@@ -12,7 +10,7 @@ ZabbixAnnotationsQueryController.templateUrl = 'datasource-zabbix/partials/annot
 
 export {
   ZabbixAPIDatasource as Datasource,
-  ZabbixConfigController as ConfigCtrl,
+  ZabbixDSConfigController as ConfigCtrl,
   ZabbixQueryController as QueryCtrl,
   ZabbixQueryOptionsController as QueryOptionsCtrl,
   ZabbixAnnotationsQueryController as AnnotationsQueryCtrl