+++ src/datasource-zabbix/query.controller.js
@@ -1,5 +1,4 @@
 import {QueryCtrl} from 'app/plugins/sdk';
-import angular from 'angular';
 import _ from 'lodash';
 import * as c from './constants';
 import * as utils from './utils';
@@ -22,17 +21,35 @@ export class ZabbixQueryController extends QueryCtrl {
     this.replaceTemplateVars = this.datasource.replaceTemplateVars;
     this.templateSrv = templateSrv;
 
-    this.editorModes = {
-      0: {value: 'num',       text: 'Metrics',     mode: c.MODE_METRICS},
-      1: {value: 'itservice', text: 'IT Services', mode: c.MODE_ITSERVICE},
-      2: {value: 'text',      text: 'Text',        mode: c.MODE_TEXT}
+    this.editorModes = [
+      {value: 'num',       text: 'Metrics',     mode: c.MODE_METRICS},
+      {value: 'text',      text: 'Text',        mode: c.MODE_TEXT},
+      {value: 'itservice', text: 'IT Services', mode: c.MODE_ITSERVICE},
+      {value: 'itemid',    text: 'Item ID',     mode: c.MODE_ITEMID}
+    ];
+
+    this.$scope.editorMode = {
+      METRICS: c.MODE_METRICS,
+      TEXT: c.MODE_TEXT,
+      ITSERVICE: c.MODE_ITSERVICE,
+      ITEMID: c.MODE_ITEMID
     };
+
+    this.slaPropertyList = [
+      {name: "Status", property: "status"},
+      {name: "SLA", property: "sla"},
+      {name: "OK time", property: "okTime"},
+      {name: "Problem time", property: "problemTime"},
+      {name: "Down time", property: "downtimeTime"}
+    ];
 
     // Map functions for bs-typeahead
     this.getGroupNames = _.bind(this.getMetricNames, this, 'groupList');
     this.getHostNames = _.bind(this.getMetricNames, this, 'hostList', true);
     this.getApplicationNames = _.bind(this.getMetricNames, this, 'appList');
     this.getItemNames = _.bind(this.getMetricNames, this, 'itemList');
+    this.getITServices = _.bind(this.getMetricNames, this, 'itServiceList');
+    this.getVariables = _.bind(this.getTemplateVariables, this);
 
     // Update metric suggestion when template variable was changed
     $rootScope.$on('template-variable-value-updated', () => this.onVariableChange());
@@ -57,14 +74,14 @@ export class ZabbixQueryController extends QueryCtrl {
 
       // Load default values
       var targetDefaults = {
-        mode: c.MODE_METRICS,
-        group: { filter: "" },
-        host: { filter: "" },
-        application: { filter: "" },
-        item: { filter: "" },
-        functions: [],
-        options: {
-          showDisabledItems: false
+        'mode': c.MODE_METRICS,
+        'group': { 'filter': "" },
+        'host': { 'filter': "" },
+        'application': { 'filter': "" },
+        'item': { 'filter': "" },
+        'functions': [],
+        'options': {
+          'showDisabledItems': false
         }
       };
       _.defaults(target, targetDefaults);
@@ -77,26 +94,11 @@ export class ZabbixQueryController extends QueryCtrl {
       if (target.mode === c.MODE_METRICS ||
           target.mode === c.MODE_TEXT) {
 
-        this.downsampleFunctionList = [
-          {name: "avg", value: "avg"},
-          {name: "min", value: "min"},
-          {name: "max", value: "max"},
-          {name: "sum", value: "sum"},
-          {name: "count", value: "count"}
-        ];
-
         this.initFilters();
       }
       else if (target.mode === c.MODE_ITSERVICE) {
-        this.slaPropertyList = [
-          {name: "Status", property: "status"},
-          {name: "SLA", property: "sla"},
-          {name: "OK time", property: "okTime"},
-          {name: "Problem time", property: "problemTime"},
-          {name: "Down time", property: "downtimeTime"}
-        ];
-        this.itserviceList = [{name: "test"}];
-        this.updateITServiceList();
+        _.defaults(target, {slaProperty: {name: "SLA", property: "sla"}});
+        this.suggestITServices();
       }
     };
 
@@ -104,7 +106,8 @@ export class ZabbixQueryController extends QueryCtrl {
   }
 
   initFilters() {
-    let itemtype = this.editorModes[this.target.mode].value;
+    let itemtype = _.find(this.editorModes, {'mode': this.target.mode});
+    itemtype = itemtype ? itemtype.value : null;
     return Promise.all([
       this.suggestGroups(),
       this.suggestHosts(),
@@ -127,6 +130,12 @@ export class ZabbixQueryController extends QueryCtrl {
     }
 
     return metrics;
+  }
+
+  getTemplateVariables() {
+    return _.map(this.templateSrv.variables, variable => {
+      return '$' + variable.name;
+    });
   }
 
   suggestGroups() {
@@ -170,6 +179,14 @@ export class ZabbixQueryController extends QueryCtrl {
     .then(items => {
       this.metric.itemList = items;
       return items;
+    });
+  }
+
+  suggestITServices() {
+    return this.zabbix.getITService()
+    .then(itservices => {
+      this.metric.itServiceList = itservices;
+      return itservices;
     });
   }
 
@@ -292,30 +309,7 @@ export class ZabbixQueryController extends QueryCtrl {
   switchEditorMode(mode) {
     this.target.mode = mode;
     this.init();
-  }
-
-  /////////////////
-  // IT Services //
-  /////////////////
-
-  /**
-   * Update list of IT services
-   */
-  updateITServiceList() {
-    this.zabbix.getITService().then((iteservices) => {
-      this.itserviceList = [];
-      this.itserviceList = this.itserviceList.concat(iteservices);
-    });
-  }
-
-  /**
-   * Call when IT service is selected.
-   */
-  selectITService() {
-    if (!_.isEqual(this.oldTarget, this.target) && _.isEmpty(this.target.errors)) {
-      this.oldTarget = angular.copy(this.target);
-      this.panelCtrl.refresh();
-    }
+    this.targetChanged();
   }
 }