+++ src/datasource-zabbix/utils.js
@@ -22,6 +22,15 @@ export function expandItemName(name, key) {
   return name;
 }
 
+export function expandItems(items) {
+  _.forEach(items, item => {
+    item.item = item.name;
+    item.name = expandItemName(item.item, item.key_);
+    return item;
+  });
+  return items;
+}
+
 function splitKeyParams(paramStr) {
   let params = [];
   let quoted = false;
@@ -93,7 +102,7 @@ function escapeMacro(macro) {
  * {group}{host.com} -> [group, host.com]
  */
 export function splitTemplateQuery(query) {
-  let splitPattern = /{[^{}]*}/g;
+  let splitPattern = /\{[^\{\}]*\}|\{\/.*\/\}/g;
   let split;
 
   if (isContainsBraces(query)) {
@@ -109,7 +118,8 @@ export function splitTemplateQuery(query) {
 }
 
 function isContainsBraces(query) {
-  return query.includes('{') && query.includes('}');
+  let bracesPattern = /^\{.+\}$/;
+  return bracesPattern.test(query);
 }
 
 // Pattern for testing regex